♻️ Implement Mixin classes for invoking and streaming LLM responses.
